use neo4j as main database

+++ src/api/response/Friend.kt
@@ -1,3 +1,3 @@
 package com.example.api.response
 
-data class Friend(val id: Int, val apps: String)
+data class Friend(val id: Long, val apps: String)
